Teste de retorno float

+++ database/migrations/2022_04_19_174804_create_properties_table.php
@@ -21,8 +21,8 @@ class CreatePropertiesTable extends Migration
             $table->enum('type', ['Casa', 'Apartamento', 'Kitnet']);
             $table->string('title');
             $table->text('description')->nullable(true);  //não obrigatório
-            $table->double('price', 8, 2);        //valor do imóvel
-            $table->double('iptu', 8, 2);         //valor do IPTU
+            $table->float('price', 8, 2);        //valor do imóvel
+            $table->float('iptu', 8, 2);         //valor do IPTU
             $table->unsignedSmallInteger('size');  //tamanho (em m²)
             $table->unsignedTinyInteger('number_rooms');      //quantidade de quartos
             $table->unsignedTinyInteger('number_bathrooms');  //quantidede de banheiros
@@ -31,10 +31,10 @@ class CreatePropertiesTable extends Migration
             $table->boolean('accepts_pet')->nullable(true);  //1 - aceita | 0 - não aceita
             $table->unsignedTinyInteger('garage')->nullable(true);    //não obrigatório
             $table->TinyInteger('apartment_floor')->nullable(true);   //não obrigatório
-            $table->double('condominium', 8, 2)->nullable(true);     //não obrigatório
+            $table->float('condominium', 8, 2)->nullable(true);     //não obrigatório
             $table->text('condominium_description')->nullable(true);  //não obrigatório
-            $table->double('fire_insurance', 8, 2)->nullable(true);  //não obrigatório
-            $table->double('service_charge', 8, 2)->nullable(true);  //não obrigatório
+            $table->float('fire_insurance', 8, 2)->nullable(true);  //não obrigatório
+            $table->float('service_charge', 8, 2)->nullable(true);  //não obrigatório
             $table->string('state');
             $table->string('cep');
             $table->string('city');
